Support more browsers
