+++ lib/Transport.js
@@ -37,7 +37,7 @@ Transport.prototype._action = function(method, url, data, callback) {
     someRequest = someRequest.send(data);
   }
   someRequest.end(function(err, payload) {
-    // console.log("=>", err, JSON.stringify(JSON.parse(payload.text), null, 2));
+    // console.log("=>", err, payload);
     if (err) {
       // console.log(err.response.text);
       if (err.status === 401) {
@@ -48,6 +48,7 @@ Transport.prototype._action = function(method, url, data, callback) {
       try {
         response = JSON.parse(err.response.text);
       } catch(e) {
+        console.error("Transport Error: " + JSON.stringify(err));
         return callback(new Error("Invalid response from server"));
       }
