+++ package.json
@@ -25,13 +25,14 @@
     "browserify": "13.0.0",
     "coveralls": "2.11.6",
     "eslint": "0.24.0",
-    "jsonapi-server": "1.2.0",
+    "jsonapi-server": "1.3.0",
     "karma": "0.13.19",
     "karma-chrome-launcher": "0.2.2",
     "karma-firefox-launcher": "0.1.7",
     "karma-mocha": "0.2.1",
     "karma-phantomjs-launcher": "0.2.3",
     "karma-sauce-launcher": "0.3.0",
+    "karma-spec-reporter": "0.0.23",
     "mocha": "2.2.5",
     "mocha-lcov-reporter": "0.0.2",
     "phantomjs": "1.9.19",
@@ -41,17 +42,13 @@
   "scripts": {
     "build": "./node_modules/browserify/bin/cmd.js lib/Client.js | ./node_modules/uglifyjs/bin/uglifyjs -cm > dist/jsonapi-client.min.js",
     "mocha": "./node_modules/mocha/bin/mocha -R spec ./test/*.js",
-
     "prekarmalocal": "node ./node_modules/jsonapi-server/example/server.js &",
     "karmalocal": "sleep 1 && ./node_modules/browserify/bin/cmd.js --ignore jsonapi-server/example/server.js test/*.js > dist/jsonapi-client-test.js && ./node_modules/karma/bin/karma start ./karma.local.conf.js",
     "postkarmalocal": "killall node",
-
     "prekarmaremote": "node ./node_modules/jsonapi-server/example/server.js &",
     "karmaremote": "sleep 1 && ./node_modules/browserify/bin/cmd.js --ignore jsonapi-server/example/server.js test/*.js > dist/jsonapi-client-test.js && ./node_modules/karma/bin/karma start ./karma.saucelabs.conf.js",
     "postkarmaremote": "killall node",
-
     "karma": "npm run karmalocal && npm run karmaremote",
-
     "test": "npm run mocha &&  npm run karma",
     "coveralls": "./node_modules/mocha/bin/mocha --require blanket --reporter mocha-lcov-reporter ./test/*.js | ./node_modules/coveralls/bin/coveralls.js",
     "coverage": "./node_modules/mocha/bin/mocha --require blanket --reporter html-cov ./test/*.js > coverage.html",
