+++ test/read.js
@@ -7,6 +7,121 @@ var client = new Client("http://localhost:16006/rest");
 
 describe("Testing jsonapi-client", function() {
 
+  context("searches for resources", function() {
+    it("resource.toJSONTree() should provide developer-friendly debug info", function(done) {
+      client.find("people", { include: "articles", filter: { lastname: "Rumbelow"}}, function(err, people) {
+        assert.equal(err, null);
+
+        var treeView = people.map(function(person) {
+          return person.toJSONTree();
+        });
+
+        assert.deepEqual(treeView, [
+          {
+            "id": "cc5cca2e-0dd8-4b95-8cfc-a11230e73116",
+            "type": "people",
+            "firstname": "Oli",
+            "lastname": "Rumbelow",
+            "email": "oliver.rumbelow@example.com",
+            "articles": [
+              {
+                "id": "de305d54-75b4-431b-adb2-eb6b9e546014",
+                "type": "articles",
+                "title": "NodeJS Best Practices",
+                "status": "published",
+                "content": "na",
+                "author": "[Circular]",
+                "tags": [
+                  {
+                    "type": "tags",
+                    "id": "7541a4de-4986-4597-81b9-cf31b6762486"
+                  }
+                ],
+                "photos": [],
+                "comments": [
+                  {
+                    "type": "comments",
+                    "id": "3f1a89c2-eb85-4799-a048-6735db24b7eb"
+                  }
+                ]
+              }
+            ],
+            "photos": null
+          }
+        ]);
+
+        done();
+      });
+    });
+
+    it("finds resources", function(done) {
+      client.find("articles", { }, function(err, people) {
+        assert.equal(err, null);
+        assert.equal(people.length, 4);
+
+        done();
+      });
+    });
+
+    it("filters resources", function(done) {
+      client.find("articles", { filter: { title: "<M" } }, function(err, people) {
+        assert.equal(err, null);
+        assert.equal(people.length, 2);
+
+        done();
+      });
+    });
+
+    it("passes back server errors", function(done) {
+      client.find("articles", { filter: { foobar: "<M" } }, function(err) {
+        assert.ok(err instanceof Error);
+        assert.equal(err.message, "\"articles do not have property foobar\"");
+
+        done();
+      });
+    });
+  });
+
+  it("gets existing resources", function() {
+    it("gets a resource", function(done) {
+      client.get("people", "ad3aa89e-9c5b-4ac9-a652-6670f9f27587", { }, function(err, people) {
+        assert.equal(err, null);
+        assert.deepEqual(people.toJSON(), {
+          id: "ad3aa89e-9c5b-4ac9-a652-6670f9f27587",
+          type: "people",
+          firstname: "Rahul",
+          lastname: "Patel",
+          email: "rahul.patel@example.com",
+          articles: undefined,
+          photos: undefined
+        });
+
+        done();
+      });
+    });
+
+    it("links included resources", function(done) {
+      client.get("people", "d850ea75-4427-4f81-8595-039990aeede5", { include: "photos" }, function(err, people) {
+        assert.equal(err, null);
+        assert.ok(people.photos);
+        assert.equal(people, people.photos[0].photographer);
+        assert.deepEqual(people.photos[0].toJSON(), {
+          id: "72695cbd-e9ef-44f6-85e0-0dbc06a269e8",
+          type: "photos",
+          title: "Penguins",
+          url: "http://www.example.com/penguins",
+          height: 220,
+          width: 60,
+          photographer: {
+            type: "people",
+            id: "d850ea75-4427-4f81-8595-039990aeede5"
+          },
+          articles: undefined
+        });
+        done();
+      });
+    });
+  });
   describe("fetches foreign resources", function() {
     var person, photo;
 
@@ -52,7 +167,7 @@ describe("Testing jsonapi-client", function() {
       var somePhoto = person.photos[0];
       assert.throws(function() {
         somePhoto.relationships("photographer").add("foobar");
-      }, /Expected Resource, got String/);
+      }, /Expected Resource, got /);
     });
 
     it("removing the linked resource works fine", function(done) {
